Moved all event handler code to their own handler classes. Removed Duplicate event handlers.
