+++ src/main/java/com/cannolicatfish/rankine/events/RankineClientEventHandler.java
@@ -2,6 +2,7 @@ package com.cannolicatfish.rankine.events;
 
 
 import com.cannolicatfish.rankine.events.handlers.client.FOVUpdateHandler;
+import com.cannolicatfish.rankine.events.handlers.client.ItemTooltipHandler;
 import com.cannolicatfish.rankine.events.handlers.client.RecipesUpdateHandler;
 import com.cannolicatfish.rankine.events.handlers.client.RenderBlockOverlayHandler;
 import com.cannolicatfish.rankine.events.handlers.client.RenderOverlayHandler;
@@ -11,6 +12,7 @@ import net.minecraftforge.client.event.FOVUpdateEvent;
 import net.minecraftforge.client.event.RecipesUpdatedEvent;
 import net.minecraftforge.client.event.RenderBlockOverlayEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
@@ -39,6 +41,12 @@ public class RankineClientEventHandler {
     @OnlyIn(Dist.CLIENT)
     public void renderOverlayBlockEvent(RenderBlockOverlayEvent event) {
         RenderBlockOverlayHandler.renderOverlayBlockEvent(event);
+    }
+
+    @SubscribeEvent
+    @OnlyIn(Dist.CLIENT)
+    public static void onTooltipCheck(ItemTooltipEvent event) {
+        ItemTooltipHandler.onTooltipCheck(event);
     }
