+++ src/main/java/com/cannolicatfish/rankine/events/handlers/common/PlayerLoginHandler.java
@@ -1,0 +1,44 @@
+package com.cannolicatfish.rankine.events.handlers.common;
+
+import com.cannolicatfish.rankine.compatibility.Patchouli;
+import com.cannolicatfish.rankine.init.Config;
+import com.cannolicatfish.rankine.items.alloys.AlloyItem;
+import com.cannolicatfish.rankine.items.alloys.IAlloyItem;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.event.entity.player.PlayerEvent;
+import vazkii.patchouli.api.PatchouliAPI;
+
+public class PlayerLoginHandler {
+    private static final String NBT_KEY = "rankine.firstjoin";
+
+    public static void onPlayerJoin( PlayerEvent.PlayerLoggedInEvent event) {
+        if (Config.GENERAL.REFRESH_ALLOYS.get()) {
+            for(int i = 0; i < event.getPlayer().inventory.getSizeInventory(); ++i) {
+                ItemStack itemstack = event.getPlayer().inventory.getStackInSlot(i);
+                if (!itemstack.isEmpty() && itemstack.getItem() instanceof IAlloyItem) {
+                    ((AlloyItem)itemstack.getItem()).setRefresh(itemstack);
+                }
+            }
+        }
+
+        if (Config.GENERAL.STARTING_BOOK.get() && !event.getPlayer().getEntityWorld().isRemote && Patchouli.isInstalled()) {
+
+            CompoundNBT data = event.getPlayer().getPersistentData();
+            CompoundNBT persistent;
+            if (!data.contains(PlayerEntity.PERSISTED_NBT_TAG)) {
+                data.put(PlayerEntity.PERSISTED_NBT_TAG, (persistent = new CompoundNBT()));
+            } else {
+                persistent = data.getCompound(PlayerEntity.PERSISTED_NBT_TAG);
+            }
+
+            if (!persistent.contains(NBT_KEY)) {
+                persistent.putBoolean(NBT_KEY, true);
+                event.getPlayer().inventory.addItemStackToInventory(PatchouliAPI.get().getBookStack(new ResourceLocation("rankine:rankine_journal")));
+            }
+        }
+    }
+
+}
