+++ src/main/java/com/cannolicatfish/rankine/events/handlers/common/WorldLoadHandler.java
@@ -1,0 +1,10 @@
+package com.cannolicatfish.rankine.events.handlers.common;
+
+import com.cannolicatfish.rankine.init.VanillaIntegration;
+import net.minecraftforge.event.world.WorldEvent;
+
+public class WorldLoadHandler {
+    public static void fuelValues( WorldEvent.Load event) {
+        VanillaIntegration.populateFuelMap();
+    }
+}
